rename/refactor or relational.py -> sqlite_store.py
